backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow
Pillow
python-multipart
python-dotenv
jinja2